requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
playwright>=1.50.0
pydantic-settings==2.4.0
gspread==6.1.2
google-auth==2.34.0
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.2
